mike - physics work now, switched to double system for smoother constant control, and optimized for 120fps (this can be changed for whatever)

+++ backend/Angler.java
@@ -7,7 +7,7 @@ public class Angler extends Player {
     }
 
     public void bite(){
-        int[] pos = this.getPosition();
+        double[] pos = this.getPosition();
         
 
     }//close attack
